COR 190913_updated the master plugin w LIVE
These files seem to have changed but I'm not sure if it's actually meaningful. I think my Git tool just automatically detected the deletion and addition of exactly the same info. It could be different but at this point, I'm just adding this here to reflect the LIVE site's vesion of the master plugin.